% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{buildGenericURL}
\alias{buildGenericURL}
\title{Build the URL for a generic data source}
\usage{
buildGenericURL(endpoint, params = list())
}
\arguments{
\item{endpoint, }{endpoint to retrieve as a character vector}

\item{params, }{list of parameters}
}
\value{
A full URL for retrieving the specified data.
}
\description{
This function is called by \code{\link{getGenericData}} to build the
URL for retrieving the specified data source.
}
\examples{
buildGenericURL("drafthistory", params = list(LeagueID = "00"))
}

