% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{buildURL}
\alias{buildURL}
\title{Builds a URL for requesting a resource.}
\usage{
buildURL(resource, params = list())
}
\arguments{
\item{resource, }{endpoint to retrieve as a character vector}

\item{params, }{list of parameters}
}
\value{
A full URL for retrieving the specified data.
}
\description{
Builds a URL for requesting a resource.
}
\examples{
buildURL("commonplayerinfo", list(playerID = 201939))
}

