% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{courtOutline}
\alias{courtOutline}
\title{Get the coordinates of lines on the court}
\usage{
courtOutline()
}
\value{
A data frame containing coordinates of line segments that, when connected,
draw the lines of an NBA court. The type column specifies which segment the point
belongs to, and the ltype column specifies whether that line segment should be solid
or dashed. The coordinates are in units of feet times 10 with the origin at the
center of the hoop, the same as that returned by the "shotchartdetail" endpoint.
}
\description{
This function gets the coordinates of the lines on the court (sideline, 
baseline, three point line, etc.) for building custom shot charts. They are expressed
in the same coordinates as those returned by the "shotchartdetail" endpoint.
}
\examples{
## plot court using base graphics
court <- courtOutline()
plot(x = NULL, xlim = c(-275, 275), ylim = c(-80, 450), xaxt = "n", yaxt = "n", ann = FALSE)
for (nm in unique(court$type)) {
   s <- court$type == nm
   points(x = court$x[s], y = court$y[s], type = "l", lty = court$ltype[s])
}
}
\references{
See \url{http://www.sportsknowhow.com/basketball/dimensions/nba-basketball-court-dimensions.html}
for more information about how an NBA court is laid out.
}

