% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftHistory.R
\docType{data}
\name{draftHistory}
\alias{draftHistory}
\title{Draft History Data}
\description{
A data frame with NBA draft picks from 1947 to 2015. This data
was retrieved with a call to \code{\link{getGenericData}} using the "drafthistory"
endpoint.
}

