% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoint data}
\description{
This data frame contains all of the endpoints from which you can retrieve data 
and their required parameters.
}
\references{
This data has been adapted from Github user seemethere's nba_py repository. See \url{https://github.com/seemethere/nba_py/wiki/stats.nba.com-Endpoint-Documentation}.
}
\keyword{data}

