% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getEndpointParams}
\alias{getEndpointParams}
\title{Get endpoint parameters to pass to \code{\link{getGenericData}}}
\usage{
getEndpointParams(endpoint)
}
\arguments{
\item{endpoint, }{a character vector containing an endpoint}
}
\value{
A character vector describing the required parameters.
}
\description{
This function returns the expected parameters for a specified endpoint. Use \code{\link{searchEndpoints}}
to search the available endpoints. Once you have an endpoint and a parameter list, a call can be made to 
\code{\link{getGenericData}}.
}
\examples{
getEndpointParams("commonplayerinfo")
}

