% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getIDMappings}
\alias{getIDMappings}
\title{Get mappings from player name and team name to PlayerID and TeamID}
\usage{
getIDMappings()
}
\value{
A list of length two containing data frames with the mappings.
}
\description{
For many endpoints, PlayerID or TeamID are required parameters. This function
retrieves a mapping from player name to PlayerID for all players in NBA history and team name 
to TeamID for all current teams. These ID's are used as parameters to pass in to 
\code{\link{getGenericData}} for several endpoints. Use \code{\link{searchIDMappings}} to search
these mappings for a specific player or team.
}
\examples{
m <- getIDMappings()
head(m$player)
head(m$team)
}

