% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getPlayerData}
\alias{getPlayerData}
\title{Retrieves a data frame containing info for a player.}
\usage{
getPlayerData(playerID)
}
\arguments{
\item{playerID, }{ID of player you want to retrieve}
}
\value{
A list of data frames containing info on the player with the specified player_id.
}
\description{
Retrieves a data frame containing info for a player.
}
\examples{
getPlayerData(playerID = 201939)
}

