% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{jsonToDF}
\alias{jsonToDF}
\title{Convert JSON response into data frame or list of data frames}
\usage{
jsonToDF(data)
}
\arguments{
\item{data, }{JSON resulting from a call to jsonlite::fromJSON}
}
\value{
A data frame or list of data frames containing the JSON data.
}
\description{
Converts the JSON response to a data frame or a list of data frames (if the
JSON has multiple row sets in it).
}

