% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApps.R
\name{launchApp}
\alias{launchApp}
\title{Launch an interactive user interface}
\usage{
launchApp(appName = "shotcharts")
}
\arguments{
\item{appName, }{The name of the application that you want to launch. 
Valid options are: "shotcharts", "draft"}
}
\description{
This function launches a shiny based application for viewing 
draft history or plotting a shot chart for a player.
}

