% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{memGetIDMappings}
\alias{memGetIDMappings}
\title{Memoised version of getIDMappings}
\usage{
memGetIDMappings()
}
\value{
A list of length two containing data frames with the mappings.
}
\description{
This function behaves exactly the same as \code{\link{getIDMappings}}
except that it is memoised. This prevents the code from making the same HTTP request
multiple times. If you call \code{\link{getIDMappings}} only once in your code, there
is no reason to use this function.
}

