% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerInfo2016.R
\docType{data}
\name{playerInfo2016}
\alias{playerInfo2016}
\title{Info for active NBA players as of March 2016}
\description{
A data frame containing info such as country, college, height,
weight, position, etc. for all active NBA players as of March 2016. This
data was retrieved with calls to \code{\link{getGenericData}} using the
"commonplayerinfo" endpoint for every active player whose ID's were found
with \code{\link{getIDMappings}}.
}

