% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNBA.R
\docType{package}
\name{rNBA}
\alias{rNBA}
\alias{rNBA-package}
\title{An R package for retrieving stats from the NBA.}
\description{
An R package for retrieving stats from the NBA.
}
\references{
All data comes from \url{http://stats.nba.com/}. If you plan to publish anything resulting from the use of this package, the data must explicity be attributed to the NBA.
}

