% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{searchIDMappings}
\alias{searchIDMappings}
\title{Search ID mappings}
\usage{
searchIDMappings(player = NA, team = NA, active = TRUE)
}
\arguments{
\item{player, }{player name to search for as a regex pattern (see \link{regex})}

\item{team, }{team name to search for as a regex pattern (see \link{regex})}

\item{active, }{flag indicating that only active players should be searched}
}
\value{
Returns a list containing one or two (depending on which of player and team is specified) 
data frame(s) containing matches.
}
\description{
For many endpoints, PlayerID or TeamID are required parameters. This function
searches the mapping returned by \code{\link{getIDMappings}} for a specified player or team,
and it returns the matches. These ID's are used as parameters to pass in to 
\code{\link{getGenericData}} for several endpoints.
}
\examples{
searchIDMappings(player = "curry")
searchIDMappings(player = "curry", team = "golden state")
}

