% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stephCurryShots.R
\docType{data}
\name{stephCurryShots}
\alias{stephCurryShots}
\title{Steph Curry Shots Data}
\description{
A data frame with all of Steph Curry's shots for a 10 game 
period in March 2016. This data is used as a default when building shot
charts with \code{\link{shotChart}}. This data was retrieved with a call
to \code{\link{getGenericData}} using the "shotchartdetail" endpoint.
}

