% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getGenericData}
\alias{getGenericData}
\title{Retrieve a generic data source}
\usage{
getGenericData(endpoint, params = list())
}
\arguments{
\item{endpoint, }{endpoint to retrieve as a character vector}

\item{params, }{list of parameters}
}
\value{
A data frame or a list of data frames containing the requested data.
}
\description{
Use \link{\code{searchEndpoints}} and \link{\code{getEndpointParams}} to learn
what arguments to pass into this function.
}
\examples{
getGenericData("commonallplayers", list(IsOnlyCurrentSeason = 0, LeagueID="00", Season="2015-16"))
}

