% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getGenericData}
\alias{getGenericData}
\title{Retrieves a generic data source}
\usage{
getGenericData(resource, params)
}
\arguments{
\item{resource, }{endpoint to retrieve as a character vector}

\item{params, }{list of parameters}
}
\value{
A data frame containing info on the player with the specified player_id
}
\description{
Retrieves a generic data source
}
\examples{
getGenericData("commonallplayers", list(IsOnlyCurrentSeason = 0, LeagueID="00", Season="2015-16"))
}

