% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getGenericData}
\alias{getGenericData}
\title{Retrieve a generic data source}
\usage{
getGenericData(endpoint, params = list())
}
\arguments{
\item{endpoint, }{endpoint to retrieve as a character vector}

\item{params, }{list of parameters}
}
\value{
A data frame or a list of data frames (if the endpoint returns multiple 
datasets) containing the requested data.
}
\description{
This function retrieves the specified data source. Use \code{\link{searchEndpoints}}
to see valid options for the endpoint parameter (these are the available datasets) and 
\code{\link{getEndpointParams}} to learn what parameters are needed to pass in as the
params argument for the chosen endpoint.
}
\examples{
player.info.endpoint <- "commonplayerinfo"
player.info.params <-list(PlayerID=201939,IsOnlyCurrentSeason=0,LeagueID="00",Season="2015-16")
player.info <- getGenericData(player.info.endpoint, player.info.params)
str(player.info)
######
game.log.endpoint <- "playergamelog"
game.log.params <- list(PlayerID = 201939, Season = "2015-16", SeasonType = "Regular+Season")
game.log <- getGenericData(game.log.endpoint, game.log.params)
str(game.log)
}

