% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getPlayerData}
\alias{getPlayerData}
\title{Retrieves a data frame containing info for a player (this is just an example.. trying to learn how packages work).}
\usage{
getPlayerData(params)
}
\arguments{
\item{playerID, }{ID of player you want to retrieve}
}
\value{
A data frame containing info on the player with the specified player_id
}
\description{
Retrieves a data frame containing info for a player (this is just an example.. trying to learn how packages work).
}
\examples{
getPlayerData(list(playerID=201939))
}

