% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{getPlayerTrackingData}
\alias{getPlayerTrackingData}
\title{Retrieve SportVU Player Tracking Data}
\usage{
getPlayerTrackingData(year, type = NULL)
}
\arguments{
\item{year, }{the season from which to retrieve data. 2014 refers to the 2014-15 season, 2015 refers to 
the 2015-16 season, etc. The earliest available year for this data is 2013 (the 2013-14 season).}

\item{type, }{a character vector containing one or more type of data to retrieve. The valid types are
"catchShootData", "defenseData", "drivesData", "passingData", "touchesData", "pullUpShootData", 
"reboundingData", "shootingData", or "speedData". The default is all of them.}
}
\value{
A data frame or list of data frames (if multiple types specified) containing the requested data.
}
\description{
This function retrieves Player Tracking stats (see the "Player Tracking" tab on 
\url{http://stats.nba.com/}.
}
\examples{
pt.data <- getPlayerTrackingData(year = 2014, c("defenseData", "catchShootData"))
str(pt.data)
}

