% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApps.R
\name{launchApp}
\alias{launchApp}
\title{Launch a UI based application to view or interaction}
\usage{
launchApp(appName = "rNBAshinyTestApp")
}
\arguments{
\item{appName, }{Application Name}
}
\description{
This function launch a shiny based application for interactive play.
}

