% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNBA.R
\docType{package}
\name{rNBA}
\alias{rNBA}
\alias{rNBA-package}
\title{An R package for retrieving stats from the NBA.}
\description{
An R package for retrieving stats from the NBA.
}
\references{
All data used and retrieved by this package comes from \url{http://stats.nba.com/}.
Please abide by their terms of use (see \url{http://www.nba.com/news/termsofuse.html}) when
publishing any work resulting from the use of this package.
}

