% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNBA.R
\docType{package}
\name{rNBA}
\alias{rNBA}
\alias{rNBA-package}
\title{An R package for retrieving stats from the NBA.}
\description{
An R package for retrieving stats from the NBA.
}
\references{
All data comes from \url{http://stats.nba.com/}.
}
\seealso{
\code{\link{getPlayerData}}
}

