% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{searchEndpoints}
\alias{searchEndpoints}
\title{Search raw endpoints that can be passed to \code{\link{getGenericData}}.}
\usage{
searchEndpoints(pattern)
}
\arguments{
\item{pattern, }{regex pattern to search for (see \link{regex})}
}
\value{
A character vector containing all available and matching endpoints.
}
\description{
This function lets you search for available endpoints. Use \code{\link{getEndpointParams}}
to get the required parameters for a specified endpoint. Once you have an endpoint and a parameter list,
a call can be made to \code{\link{getGenericData}}.
}
\examples{
searchEndpoints(".")
searchEndpoints("player")
}

